fix: Only initialise logger if it doesn't exist
Trying to create a new logger with a name that already exists results in
an error.  Thus, first check if there already is a logger and only
create a new one if not.
This should also allow custom configuration of the logger by creating
and configuring one before instantiating the driver class.