More generic `create_solo12_backend()`
Make `create_solo12_backend()` more generic by providing the driver as
argument and adding an option whether to use `MonitoredRobotDriver` or
not.  Use this in the other "create backend"-functions to avoid
redundancy.

Keep the old version for convenience but rename it to
`create_real_solo12_backend()` to be more explicit.