fix: Set NO_EXTRAS for pybind11 when using GCC
For some unresolved reason, adding Python bindings for
PyBulletSolo12Driver causes an internal compiler error of GCC on some
machines (happens on GitHub workflows, I could also reproduce it on a
Dell PC but not on my ThinkPad...).  It happens with both GCC 9 and 10.

I could not figure out what exactly is causing the error but it is
related to LTO.  Setting NO_EXTRAS to disable this kind of optimisation
avoids the error.  This is not a good permanent solution but for now I
don't see a viable alternative (hopefully this resolves itself with
newer versions of GCC).

Only set the flag when GCC is detected as clang does not suffer from
that issue (so building with clang is a way to avoid the error but keep
the optimisations).